+++ limitador/src/storage/redis/redis_async.rs
@@ -37,7 +37,7 @@ impl AsyncCounterStorage for AsyncRedisStorage {
             .await?
         {
             Some(val) => Ok(val - delta >= 0),
-            None => Ok(counter.max_value() - delta >= 0),
+            None => Ok((((counter.max_value() as i128) - delta as i128) as i64) >= 0),
         }
     }
 
@@ -80,7 +80,7 @@ impl AsyncCounterStorage for AsyncRedisStorage {
                     }
                 }
                 None => {
-                    if counter.max_value() - delta < 0 {
+                    if (((counter.max_value() as i128) - delta as i128) as i64) < 0 {
                         return Ok(Authorization::Limited(
                             counter.limit().name().map(|n| n.to_owned()),
                         ));
